chore(middleware): cleanup files and names
